Move duplicable logic inside of schema form
